refactor(app): split AppMainWorkspaceInner into 12 focused hooks and ThreadItem
Extract ~2000 lines of self-contained logic from src/App.tsx (6625 -> 4598
lines, -30%) into single-responsibility modules under src/hooks/* and
src/app/*. AppMainWorkspaceInner shrinks from ~300 to ~150 declarations
without changing any external behavior, IPC ordering, effect timing or
ref initialization.

New hooks:
- useResizeRails: sidebar/editor-terminal drag handles + persistence
- useUiZoom: UI zoom, fullscreen, window menu
- useEditCommands: Monaco/DOM edit commands + clipboard
- useLayoutWindows: Agent/Editor layout window switching
- useWizardSends: skill/rule/subagent wizard sends (DRY'd via runWizardSend factory)
- useMessagesScroll: message viewport scroll, sticky-bottom, jump button
- useAgentPatchActions: dismiss/keep/revert/apply agent diff actions
- useThreadActions: new/select/delete/rename/navigate threads + Ctrl+N
- useComposerAttachments: composer attachments, drop, skill, MCP toggle
- useSettingsPersistence: settings IPC persistence + window availability
- useWorkspaceExplorerActions: workspace explorer right-click actions
- useAppShellSlices: top-level App's 5 slices + theme-cascade effects

New component:
- app/ThreadItem.tsx: extracted from renderThreadItem (memoized leaf)

openAgentSidebarFilePreview, onEditorTerminalSessionExit, onSendRef and
onAbortRef bodies stay in App.tsx because their dependency graphs are
too wide to externalize cleanly.

Verified with `npx tsc --noEmit` and `npm test` (377/377 passing) after
each phase. No linter errors in any new file.

Made-with: Cursor
